+++ package.json
@@ -12,7 +12,12 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/core": "^7.16.0",
+    "@babel/eslint-parser": "^7.16.3",
+    "babel-eslint": "^10.1.0",
     "css-loader": "^6.5.1",
+    "eslint": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
     "webpack": "^5.64.4",
@@ -20,6 +25,7 @@
     "webpack-dev-server": "^4.6.0"
   },
   "dependencies": {
+    "eslint-config-airbnb": "^19.0.2",
     "lodash": "^4.17.21"
   }
 }
