+++ src/fakeList.js
@@ -1,0 +1,19 @@
+const fakeList = [
+  {
+    description: 'Wash dishes',
+    completed: false,
+    index: 1,
+  },
+  {
+    description: 'Buy Groceries',
+    completed: false,
+    index: 2,
+  },
+  {
+    description: 'Walk the dog',
+    completed: false,
+    index: 3,
+  },
+];
+
+export default fakeList;
